Bug 25186: Fix accordion sections height in columns configuration page
The height for each section was the height of the highest section.
Now the height is set to the length of the section.

Test plan:
Go to Administration > Column configuration
Open the sections and confirm that the height fits its content.

QA: note that I guess "autoHeight: false" was meant to achieve that,
but it does not exist in the API of the plugin.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Josef Moravec <josef.moravec@gmail.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Joy Nelson <joy@bywatersolutions.com>
(cherry picked from commit 8641888fa3c2dd460a1cf62855dc65bc1448682c)

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>