Bug 25234: Update OPAC search results pagination with aria labels
This patch updates the include file which generates OPAC search results
pagination so that it has better semantic markup and correct aria
labels.

To test, apply the patch and do a search in the OPAC which will return
multiple pages of search results.

View the source to confirm the markup changes:

 - <nav> instead of <div> surrounding the list of links, with a
   corresponding aria label
 - Navigation links should have labels indicating which page in the
   results they point to.
 - The current link should have "aria-disabled," and "aria-current" set.
 - Numbered links should have aria labels that indicate their number.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>