Bug 25271: Add floating toolbar to OPAC cart
This patch adds JavaScript to the OPAC cart view so that the toolbar at
the top of the display "floats" as you scroll down.

The patch also makes a minor correction to the OPAC SCSS affecting the
display of the buttons in the cart toolbar.

To test, apply the patch and rebuild the OPAC CSS
(https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client).

 - Add some items to a cart in the OPAC.
 - Open the cart and confirm that the toolbar sticks to the top of the
   window as you scroll down.
 - Confirm that it works in the "More details" view.
 - Confirm that other JavaScript-based operations are not affected:
   table sorting and all the various toolbar buttons.

Signed-off-by: Frédéric Demians <f.demians@tamil.fr>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>