Bug 25276: Correct hover style of list share button in the OPAC
This patch adds missing "btn" classes to the OPAC share button so that
its style is consistent with similar controls.

The patch also makes some general changes to the OPAC CSS to make sure
link color and hover color are applied with enough specificity. This
corrects the hover color of the share button but should not change any
other existing style.

To test you should have the OpacAllowSharingPrivateLists preference
enabled.
 - Rebuild the OPAC CSS
   (https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client).
 - Log in to the OPAC as a user with one or more private lists.
 - Go to Lists -> Your lists.
 - In the list of lists there should be a "Share" link for each list.
   Hovering your mouse pointer over the link should change the style in
   the same way the "Edit" link does.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>