Bug 13388: (follow-up) Remove display of email address
Since the email address configured for a library may not be the email
address libraries want to use as a contact email address the email
should not be displayed by default.

This patch removes display of the email address. Libraries may choose to
add the information to the opac_info column, so there is already a
built-in alternative.

To test, apply the patch and view the detail page for a library in your
system which has an email address configured. The email address should
not display on the page.

Signed-off-by: Victor Grousset/tuxayo <victor@tuxayo.net>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>